Module spirit_tokio::handlers

Various Transformations for working with resources in async contexts.

Oftentimes, one needs to transform some resource into a future and spawn it into the runtime. While possible to do manually, the types here might be a bit more comfortable.

Structs

Acceptor
A plumbing type for PerConnection.

PerConnection
A Transformation that creates a new future from each accepted connection.

PerConnectionInit
A more flexible (and mind-bending) version of PerConnection.

ToFuture
A Transformation to take a resource, turn it into a future and install it.

ToFutureUnconfigured
A Transformation to take a resource, turn it into a future and install it.