Struct spirit_tokio::handlers::PerConnectionInit

pub struct PerConnectionInit<F>(pub F);

A more flexible (and mind-bending) version of PerConnection.

The PerConnection applies a closure to each accepted connection. To share the closure between all listeners, the closure is cloned.

This version's closure is higher-level closure. It is called once for each listener to produce a per-listener closure to handle its connections. In effect, it allows for custom „cloning“ of the closure.