Struct spirit_tokio::installer::FutureInstaller

pub struct FutureInstaller;

An installer able to install (an uninstall) long-running futures.

This is an installer that can be used with Fragment that produce futures.

If the spirit does not contain a Tokio runtime yet, one (the default one) will be added as a singleton. Note that this has effect on how spirit manages lifetime of the application.

An uninstallation is performed by dropping canceling the future.

This works with both concrete types (implementing Future<Output = ()> + Send) and boxed ones. Note that the boxed ones are Pin<Box<dyn Future<Output = ()> + Send>>, created by Box::pin.

See the crate level examples for details how to use (the installer is used only as the associated type in the fragment implementation).