Module spirit_tokio::net

Autoconfiguration of network primitives of tokio

Available only with the net feature. This contains several Fragments for configuring network primitives.

Modules

limits
Handling limits and errors on listening sockets.

unix
Support for unix domain sockets.

Structs

AcceptFuture
A plumbing type, return value of Accept::accept.

ConfiguredListener
A wrapper that applies configuration to each accepted connection.

Listen
A description of listening interface and port.

TcpConfig
An implementation of the StreamConfig trait to configure TCP connections.

TcpListen
A configuration fragment of a TCP listening socket.

UdpListen
A configuration fragment describing a bound UDP socket.

Enums

MaybeDuration
Configuration that can be unset, explicitly turned off or set to a duration.

Traits

Accept
Abstraction over endpoints that accept connections.

StreamConfig
Abstracts over a configuration subfragment that applies further settings to an already accepted stream.

Type Definitions

MinimalTcpListen
A TcpListen with all parameters set to Empty.

TcpListenWithLimits
Convenience type alias for configuration fragment for TCP listening socket with handling of accept errors and limiting number of current connections.