Struct spirit_tokio::net::Listen

#[non_exhaustive]pub struct Listen {
    pub port: u16,
    pub host: IpAddr,
    pub reuse_addr: Option<bool>,
    pub reuse_port: Option<bool>,
    pub only_v6: Option<bool>,
    pub backlog: u32,
    pub ttl: Option<u32>,
}

A description of listening interface and port.

This can be used as part of configuration to describe a socket.

Note that the Default implementation is there to provide something on creation of Spirit, but isn't very useful. It'll listen on a OS-assigned port on all the interfaces.

This is used as the base of the TcpListen and UdpListen configuration fragments.

Configuration options

port (mandatory)

host (optional, if not present, :: is used)

reuse-addr (optional, boolean, if not present the OS default is used)

reuse-port (optional, boolean, if not present the OS default is used, does something only on unix).

only-v6 (optional, boolean, if not present the OS default is used, does nothing for IPv4 sockets).

backlog (optional, number of waiting connections to be accepted in the OS queue, defaults to 128)

ttl (TTL of the listening/UDP socket).

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.
port: u16

The port to bind to.

host: IpAddr

The interface to bind to.

Defaults to :: (IPv6 any).

reuse_addr: Option<bool>

The SO_REUSEADDR socket option.

Usually, the OS reserves the host-port pair for a short time after it has been released, so leftovers of old connections don't confuse the new application. The SO_REUSEADDR option asks the OS not to do this reservation.

If not set, it is left on the OS default (which is likely off = doing the reservation).

reuse_port: Option<bool>

The SO_REUSEPORT socket option.

Setting this to true allows multiple applications to simultaneously bind to the same port.

If not set, it is left on the OS default (which is likely off).

This option is not available on Windows and has no effect there.

only_v6: Option<bool>

The IP_V6ONLY option.

This has no effect on IPv4 sockets.

On IPv6 sockets, this restricts the socket to sending and receiving only IPv6 packets. This allows another socket to bind the same port on IPv4. If it is set to false, the socket can communicate with both IPv6 and IPv4 endpoints under some circumstances.

Due to platform differences, the generally accepted best practice is to bind IPv4 and IPv6 as two separate sockets, the IPv6 one with setting IP_V6ONLY explicitly to true.

If not set, it is left on the OS default (which differs from OS to OS).

backlog: u32

The accepting backlog.

Has no effect for UDP sockets.

This specifies how many connections can wait in the kernel before being accepted. If more than this limit are queued, the kernel starts refusing them with connection reset packets.

The default is 128.

ttl: Option<u32>

The TTL field of IP packets sent from this socket.

If not set, it defaults to the OS value.

Implementations

impl Listen

pub fn create_tcp(&self) -> Result<StdTcpListener, AnyError>

Creates a TCP socket described by the loaded configuration.

This is the synchronous socket from standard library. See TcpListener::from_std.

pub fn create_udp(&self) -> Result<StdUdpSocket, AnyError>

Creates a UDP socket described by the loaded configuration.

This is the synchronous socket from standard library. See UdpSocket::from_std.