Struct spirit_tokio::net::TcpListen

#[non_exhaustive]pub struct TcpListen<ExtraCfg = Empty, TcpStreamConfigure = TcpConfig> {
    pub listen: Listen,
    pub tcp_config: TcpStreamConfigure,
    pub extra_cfg: ExtraCfg,
}

A configuration fragment of a TCP listening socket.

The Fragment creates a TcpListener (wrapped in ConfiguredListener). It can be handled directly, or through Pipelines.

Note that this stream sometimes returns errors „in the middle“, but most stream consumers terminate on the first error. You might be interested in the WithListenLimits wrapper to handle that automatically (or, see the TcpListenWithLimits for a convenient type alias).

Type parametes

ExtraCfg: Additional application specific configuration that can be extracted and used by the application code. This doesn't influence the socket created in any way.

TcpStreamConfigure: Further configuration for the accepted TCP streams in the form of StreamConfig

Fields

The configuration fields are pooled from all three type parameters above and from the Listen configuration fragment.

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.
listen: Listen

The actual listener socket address.

tcp_config: TcpStreamConfigure

Configuration to be applied to the accepted connections.

extra_cfg: ExtraCfg

Arbitrary application specific configuration that doesn't influence the sockets created.