Trait spirit_tokio::net::Accept

pub trait Accept: Send + Sync + 'static {
    type Connection: Send + Sync + 'static;
    fn poll_accept(
        &mut self, 
        ctx: &mut Context<'_>
    ) -> Poll<Result<Self::Connection, IoError>>;

    fn accept(&mut self) -> AcceptFuture<'_, Self> { ... }
}

Abstraction over endpoints that accept connections.

Associated Types

type Connection: Send + Sync + 'static

The type of the accepted connection.

Loading content...
Required methods

fn poll_accept(
&mut self,
ctx: &mut Context<'_>
) -> Poll<Result<Self::Connection, IoError>>

Poll for availability of the next connection.

In case there's none ready, it returns Poll::Pending and the current task will be notified by a waker.

Loading content...
Provided methods

fn accept(&mut self) -> AcceptFuture<'_, Self>

Accept the next connection.

Returns a future that will yield the next connection.

Loading content...