Trait spirit_tokio::net::IntoIncoming

pub trait IntoIncoming: Send + Sync + 'static {
    type Connection: Send + Sync + 'static;
    type Incoming: Stream<Item = Self::Connection, Error = IoError> + Send + Sync + 'static;
    fn into_incoming(self) -> Self::Incoming;
}

Abstraction over endpoints that accept connections.

The TcpListener has the incoming method that produces a stream of incoming connections. This abstracts over types with similar functionality ‒ either wrapped TcpListener, UnixListener on unix, types that provide encryption on top of these, etc.

Associated Types

type Connection: Send + Sync + 'static

The type of one accepted connection.

type Incoming: Stream<Item = Self::Connection, Error = IoError> + Send + Sync + 'static

The stream produced.

Loading content...
Required methods

fn into_incoming(self) -> Self::Incoming

Turns the given resource into the stream of incoming connections.

Loading content...