Module spirit_tokio::net::unix

Support for unix domain sockets.

This is equivalent to the configuration fragments in the net module, but for unix domain sockets. This is available only on unix platforms.

If you want to accept both normal (IP) sockets and unix domain sockets as part of the configuration, you can use the Either enum.

Structs

DatagramListen
A Fragment for unix datagram sockets

Listen
Configuration of where to bind a unix domain socket.

UnixListen
A listener for unix domain stream sockets.

Type Definitions

MinimalUnixListen
Type alias for UnixListen without any unnecessary configuration options.

UnixConfig
Additional configuration for unix domain stream sockets.

UnixListenWithLimits
Wrapped UnixListen that limits the number of concurrent connections and handles some recoverable errors.